chore: #57 - Response 모델 분리
